Enable serialization of ppoprf structs.
Enable serde and use it to support serialization of the public
function types needed to communicate with the ppoprf eval function
so they can be used to implement a randomness server.

This uses the built-in curve25519-dalek serialization implementation,
which turns points and keys into `Vec<u8>` instead of a base64
`String` which would be better for a web api. This is helpful
for getting things going, however.