Add a background puncture task to the example server.
Wrap the ppoprf::Server struct in a RwLock so it can be used
through shared read-only access by the normal request handling
threads, but be locked for mutable access when we need to.

Spawn a background task that waits asynchronously for the epoch
interval, defined as 5 seconds here for testing, and then invoke
puncture.

Launching `cargo run --example client` in the first five seconds
completes as before. After that the background task punctures
the single test md, and subsequent post requests will return an
error from the caught panic using a no-longer-accessible metadata
tag.