Demonstration randomness server (#13)
* Stub out a randomness server example.

Start an example of a PPOPRF randomness server using the actix-web
framework. This just responds to GET with a short help string and
doesn't expose any functions.

Verify the test server with:

    cargo run --example server
    [in another terminal] curl localhost:8080
    Press [ctrl-C] to exit

* Add logging to randomness server.

This makes testing easier.

* Enable serialization of ppoprf structs.

Enable serde and use it to support serialization of the public
function types needed to communicate with the ppoprf eval function
so they can be used to implement a randomness server.

This uses the built-in curve25519-dalek serialization implementation,
which turns points and keys into `Vec<u8>` instead of a base64
`String` which would be better for a web api. This is helpful
for getting things going, however.

* Evaluate points sent to the randomness server.

Accept and return a list of `CompressedRistretto` points after
passing them through the PPOPRF evaluation function.

* Document how to verify the example randomness server.

* cargo fmt.

Reformat recent server example changes.

* Stub out an example randomness client.

Just pings the server with a GET request to verify it's listening.

The idea is to add test requests for a randomness sample and the
other exposed APIs in a more automatable way than curl command lines.

* Send the ristretto base point from the example client.

Test the `eval` wrapper in the example server by requesting
random mixing of a sample point. This does the same thing
as the `curl` example in the example server docs.

* Use the basepoint contant in the example client.

Instead of duplicating the compressed ristretto base point for
the test query in an awkward serialization, use the constant
exported by the curve25519-dalek crate, which we already link.

Serialize the query to json from a native struct for better
readability.

* Add a background puncture task to the example server.

Wrap the ppoprf::Server struct in a RwLock so it can be used
through shared read-only access by the normal request handling
threads, but be locked for mutable access when we need to.

Spawn a background task that waits asynchronously for the epoch
interval, defined as 5 seconds here for testing, and then invoke
puncture.

Launching `cargo run --example client` in the first five seconds
completes as before. After that the background task punctures
the single test md, and subsequent post requests will return an
error from the caught panic using a no-longer-accessible metadata
tag.

* Add env config to server example, support for multiple metadata tags/punctures

* Ristretto serialization in base64

* Log default config values, if config not provided via env vars

* Improve example server logging.

Report initial mds vector as well as how frequently and which md
is punctured by the background task.

* Distinguish reqwest client in the ppoprf example client.

To avoid confusion between ppoprf::Client and
reqwuest::blocking::Client, import the later under a more distinct
name and use it consistently instead of the mixture of `get` and
`post` methods.

* ppoprf: Fix tests.

Propagate the Point wrapper struct to other methods so the
tests compile again.

* ppoprf: Make more types serializable.

Mark both Point and Evaluation as both serializable and deserializable
so they can be used by a randomness client as well as a server.

Also make the members of the Evaluation public for non-serialized
access.

* ppoprf: Blind points the example client sends.

Demonstrate the blind/unblind steps a client would take to mix
randomness into the hash of a real Star protocol message instead
of always using the base point for testing.

Demonstrating proof verification and finalization requires
coordinating the epoch md with the server, probably by exposing
an endpoint reporting the output of get_public_keys(), and a
schedule of when different md values are used.

* ppoprf: cargo fmt

Reformat the recently changed code to pass the ci lint.

Co-authored-by: Darnell Andries <dandries@brave.com>