Update ppoprf randomness server to actix-web v4. (#22)
Bump the actix-web version to pick up the fix for RUSTSEC-2021-0124
reported by `cargo audit` against this dev-dependency. This is a
new major release, so some minor api changes are needed.

This clones the server state into a web::Data wrapper which is
itself clone in the per-thread http responder factory closure.
This is proabably redundandant with both structs having an Arc;
if so, passing the web::Data wrapper to the background task would
be more efficient.

In practice, this arrangement seems to work under local testing
with `cargo run --example client` and `baton`.