sharks: Prefer rebinding to mutable variable.
In the documentation example, we try to recover a key from
both sufficient and insufficient number of shares. The example
used a `let mut secret` to store the result of both operations,
however in Rust it's generally better to use a new immutable
variable whenever possible.

Rebinding the previous variable name is considered acceptable
when it's a refinement of the same concept.