Some small fixups (#31)
* Fix up documentation whitespace.

Remove trailing whitespace and apply a final line-ending character
to SECURTY.md and LICENSE. Some editors lint about or silently fix
these nits, so it makes for less-noisy patches if they're uniformly
applied to files in the repo.

* ppoprf: Fix comment typo

* sharks: Prefer rebinding to mutable variable.

In the documentation example, we try to recover a key from
both sufficient and insufficient number of shares. The example
used a `let mut secret` to store the result of both operations,
however in Rust it's generally better to use a new immutable
variable whenever possible.

Rebinding the previous variable name is considered acceptable
when it's a refinement of the same concept.

* ppoprf: don't request zeroize alloc feature

This feature is on by default, so it's not necessary to activate
it explicitly.