Update fuzz targets.
These need to be updated for the API change from `u8` to `u32`
for the threshold. However, requesting 2^32 quickly crashes
with a memory allocation failure, so passing the full value
isn't immediately helpful for fuzzing. Here we continue to
use an 8-bit source value, just converting to u32 at the
call site.

A u16 might be more realistic.