sta-rs: avoid late initialization in benchmark.
Refactor the the `get_messages` helper to address a clippy issue.

It would like us to assign directly to `messages` but is confused
by the feature-gated `star2` code which makes the obvious fix
not work. Instead, swap the arms of the conditional so the simple
default flow is at the top and use a temporary in the more complex
branch.

Note that the `star2` version doesn't compile any more; it needs
to be updated for recent ppoprf api changes.