ppoprf: Disable cookies in actix-web examples
Build actix-web without the `cookies` feature enabled. This depends
on time v0.3.14 which as an minimum-supported Rust version of 1.59.0.

This is one revision past our MSRV, and the MSRV of the Brave browser
whose builds we're currently targetting. Since we don't need cookie
support for the ppoprf web service examples, disabling the them is
the simplest response.

This also disables compression, which was doing no harm, but is
unlikely to help much with the current data formats, so I'd rather
have a shorter feature list.