ppoprf: remove dotenvy
This was used in the example ppoprf randomness server to simplify
configuration for testing. The latest release requires a newer
rust release than our mimimum-supported version and is breaking
integration tests.

Since it's not essential remove it for now. It can be restored
if we can agree on a compatible support floor with upstream. In
the meantime, configuration variables must be set directly in
the environment.