Remove unused subtle dependency.
We depend transitively on the `subtle` crate through the `ff` and
`strobe-rs` (sharks) and `curve25519-dalek` (ppoprf) crates. And
while we deal like `CtOption` as a return type, we don't need
direct access to that crate's traits or symbols.

As such, the reference in sharks/Cargo.toml is unused; the crate
is never referenced. It can be removed with no change in behaviour.