sharks: Add a test to verify FIELD_ELEMENT_LEN
Check that the FIELD_ELEMENT_LEN constant we export actually
matches the number of bytes used by our `Fp` type. We could
set the constant directly from `size_of`, and I think sizing
of byte arrays is consistent in Rust, but it feels safer to
have a fixed constant since other code has hard-coded this
in the past.

This was we at least get some feedback if something changes.