adss: Use OsRng
Use the system random number generator instead of the thead-local
one. This is slower, but may address concerns about thread-safety
over ffi. Tie to the same `rand_core` minimum version as the the
other crates in the repo.

In theory this is slower because of the syscall overhead, but
I couldn't measure a difference with `cargo bench` in the
sharks subdirectory.