Vendor strobe-rng inside the ppoprf crate
The `strobe-rng` crate is only a single page of code, which is
very light for publishing as an independent crate. Ideally it
could be added as an optional feature on the upstream `strobe-rs`
crate.

In the meantime, to allow publishing our crates without that
dependency, convert the code to a module within the ppoprf
crate. This duplicates the implementation, which is an extra
maintanence burden, so this should be removed in favor of
an upstream feature or a published helper crate in the future.