star: Specify explicit versions for path dependencies
When publishing, cargo deletes the `path` key from dependencies,
so the referenced crate is resolved against versions published
to the registry instead of the local tree.

That means there needs to be a version specification. Since
rust 1.64, cargo is supposed to infer the version for path-only
dependencies from the version declared for the local copy.
This worked as expected when we published `adss` which depends
on `sharks` but for some reason is failing here.

Work around the issue by specifying a useful minimum version
explicitly in the star Cargo.toml.