sharks: Fix no-std benchmarks
Likewise update the code to restore `cargo bench --no-default-features`
by disabling benchmark functions that require the `std` feature.

The main issue here is the `dealer` method. We could replace this
with a local seedable_rng, and a call to `dealer_gen` but since
the former is just a wrapper around the later, I don't think we
gain much by benchmarking one way or the other. I just wanted
the build not to fail.