ppoprf: Add method to export the oprf key
Support private key extraction so multiple instances of the same
ppoprf::Server can be created, for example to support a cluster
of indistinguishable instances to scale throughput.

Adds a new `ServerPrivateKey` which can be extracted and converted
to a byte slice. Since it just wraps `RistrettoScalar` I picked this
over serde traits.

Also add a new `ppoprf::Server` contstructor which accepts a
key rather than generating a new one.

NB: This only represents the `oprf_key` field of the `Server`
struct. The `GGM` instance has its own random internal state,
so this commit IS NOT SUFFICIENT for repeatable behavior.
A more complex key structure is necessary.