Implement Arbitrary for Share
Avoid false positives from `cargo +nightly fuzz run recover` where
it constructs `Share` values that don't respect type invariants
and so are impossible to input through safe code.

Re-use the TryFrom<&[u8]> to create a share from an
Arbitrary sequence of bytes. This picks an artitrary
length; the the number of `y` points from that length
from the size of the submitted byte slice.

Modelled on the `&str` impl in the `arbitrary` crate.