github-actions: pin dtolnay/rust-toolchain revision
Use a fixed revision instead of the default branch so we get
a chance to review any changes.

This action defines branches and tags aligned with each rust
release, so one can use `@stable` or `@1.56.1` to specify
action script and toolchain version on one line. However this
is incompatible with pinning the action script itself.

The instructions say to use `@master` in combination with
a `with` clause specifying the toolchain version like we do
to allow a matrix expansion. But if I've understood the current
code correctly, that's not necessary because the version is
just passed to `rustup` after some filtering.