tests: Replace explicit counter with a zip
Instead of using an explicit mutable counter to generate an
enumeration value for use as AssociatedData, zip the iterator
with an open range. This saves a line and follows the functional
pattern the code is already using.

Note that we can't use the more obvious `enumerate` method here
because that yields an explicit `usize` counter value and we want
the compiler to infer the expected `u8` type.